pyproject.toml
[project]
name = "smartpath-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "multipart>=1.3.0",
    "openai>=2.16.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
]
